refactor(ibooking): move sit logic to /providers
